feat: added notifications in legacy discussion experience
