+++ lms/djangoapps/discussion/rest_api/tests/test_utils.py
@@ -3,16 +3,13 @@ Tests for Discussion REST API utils.
 """
 
 from datetime import datetime, timedelta
-from unittest.mock import Mock
 
 import ddt
-from django.conf import settings
-from httpretty import httpretty
 from pytz import UTC
 import unittest
 from common.djangoapps.student.roles import CourseStaffRole, CourseInstructorRole
 from lms.djangoapps.discussion.django_comment_client.tests.utils import ForumsEnableMixin
-from lms.djangoapps.discussion.rest_api.tests.utils import CommentsServiceMockMixin, ThreadMock
+from lms.djangoapps.discussion.rest_api.tests.utils import CommentsServiceMockMixin
 from openedx.core.djangoapps.discussions.models import PostingRestriction, DiscussionsConfiguration
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
 from xmodule.modulestore.tests.factories import CourseFactory
@@ -26,9 +23,8 @@ from lms.djangoapps.discussion.rest_api.utils import (
     get_moderator_users_list,
     get_archived_topics,
     remove_empty_sequentials,
-    send_response_notifications, is_posting_allowed
+    is_posting_allowed
 )
-from openedx_events.learning.signals import USER_NOTIFICATION_REQUESTED
 
 
 class DiscussionAPIUtilsTestCase(ModuleStoreTestCase):
@@ -163,185 +159,6 @@ class TestRemoveEmptySequentials(unittest.TestCase):
         ]
         result = remove_empty_sequentials(data)
         self.assertEqual(result, expected_output)
-
-
-def _get_mfe_url(course_id, post_id):
-    """
-    get discussions mfe url to specific post.
-    """
-    return f"{settings.DISCUSSIONS_MICROFRONTEND_URL}/{str(course_id)}/posts/{post_id}"
-
-
-class TestSendResponseNotifications(ForumsEnableMixin, CommentsServiceMockMixin, ModuleStoreTestCase):
-    """
-    Test for the send_response_notifications function
-    """
-
-    def setUp(self):
-        super().setUp()
-        httpretty.reset()
-        httpretty.enable()
-
-        self.user_1 = UserFactory.create()
-        self.user_2 = UserFactory.create()
-        self.user_3 = UserFactory.create()
-        self.thread = ThreadMock(thread_id=1, creator=self.user_1, title='test thread')
-        self.thread_2 = ThreadMock(thread_id=2, creator=self.user_2, title='test thread 2')
-        self.thread_3 = ThreadMock(thread_id=2, creator=self.user_1, title='test thread 3')
-        self.course = CourseFactory.create()
-
-    def test_send_notification_to_thread_creator(self):
-        """
-        Test that the notification is sent to the thread creator
-        """
-        handler = Mock()
-        USER_NOTIFICATION_REQUESTED.connect(handler)
-
-        # Post the form or do what it takes to send the signal
-
-        send_response_notifications(self.thread, self.course, self.user_2, parent_id=None)
-        self.assertEqual(handler.call_count, 1)
-        args = handler.call_args[1]['notification_data']
-        self.assertEqual([int(user_id) for user_id in args.user_ids], [self.user_1.id])
-        self.assertEqual(args.notification_type, 'new_response')
-        expected_context = {
-            'replier_name': self.user_2.username,
-            'post_title': 'test thread',
-            'course_name': self.course.display_name,
-        }
-        self.assertDictEqual(args.context, expected_context)
-        self.assertEqual(
-            args.content_url,
-            _get_mfe_url(self.course.id, self.thread.id)
-        )
-        self.assertEqual(args.app_name, 'discussion')
-
-    def test_send_notification_to_parent_threads(self):
-        """
-        Test that the notification signal is sent to the parent response creator and
-        parent thread creator, it checks signal is sent with correct arguments for both
-        types of notifications.
-        """
-        handler = Mock()
-        USER_NOTIFICATION_REQUESTED.connect(handler)
-
-        self.register_get_comment_response({
-            'id': self.thread_2.id,
-            'thread_id': self.thread.id,
-            'user_id': self.thread_2.user_id
-        })
-
-        send_response_notifications(self.thread, self.course, self.user_3, parent_id=self.thread_2.id)
-        # check if 2 call are made to the handler i.e. one for the response creator and one for the thread creator
-        self.assertEqual(handler.call_count, 2)
-
-        # check if the notification is sent to the thread creator
-        args_comment = handler.call_args_list[0][1]['notification_data']
-        args_comment_on_response = handler.call_args_list[1][1]['notification_data']
-        self.assertEqual([int(user_id) for user_id in args_comment.user_ids], [self.user_1.id])
-        self.assertEqual(args_comment.notification_type, 'new_comment')
-        expected_context = {
-            'replier_name': self.user_3.username,
-            'post_title': self.thread.title,
-            'author_name': 'dummy\'s',
-            'course_name': self.course.display_name,
-        }
-        self.assertDictEqual(args_comment.context, expected_context)
-        self.assertEqual(
-            args_comment.content_url,
-            _get_mfe_url(self.course.id, self.thread.id)
-        )
-        self.assertEqual(args_comment.app_name, 'discussion')
-
-        # check if the notification is sent to the parent response creator
-        self.assertEqual([int(user_id) for user_id in args_comment_on_response.user_ids], [self.user_2.id])
-        self.assertEqual(args_comment_on_response.notification_type, 'new_comment_on_response')
-        expected_context = {
-            'replier_name': self.user_3.username,
-            'post_title': self.thread.title,
-            'course_name': self.course.display_name,
-        }
-        self.assertDictEqual(args_comment_on_response.context, expected_context)
-        self.assertEqual(
-            args_comment_on_response.content_url,
-            _get_mfe_url(self.course.id, self.thread.id)
-        )
-        self.assertEqual(args_comment_on_response.app_name, 'discussion')
-
-    def test_no_signal_on_creators_own_thread(self):
-        """
-        Makes sure that no signal is emitted if user creates response on
-        their own thread.
-        """
-        handler = Mock()
-        USER_NOTIFICATION_REQUESTED.connect(handler)
-        send_response_notifications(self.thread, self.course, self.user_1, parent_id=None)
-        self.assertEqual(handler.call_count, 0)
-
-    def test_comment_creators_own_response(self):
-        """
-        Check incase post author and response auther is same only send
-        new comment signal , with your as author_name.
-        """
-        handler = Mock()
-        USER_NOTIFICATION_REQUESTED.connect(handler)
-
-        self.register_get_comment_response({
-            'id': self.thread_3.id,
-            'thread_id': self.thread.id,
-            'user_id': self.thread_3.user_id
-        })
-
-        send_response_notifications(self.thread, self.course, self.user_3, parent_id=self.thread_2.id)
-        # check if 1 call is made to the handler i.e. for the thread creator
-        self.assertEqual(handler.call_count, 1)
-
-        # check if the notification is sent to the thread creator
-        args_comment = handler.call_args_list[0][1]['notification_data']
-        self.assertEqual(args_comment.user_ids, [self.user_1.id])
-        self.assertEqual(args_comment.notification_type, 'new_comment')
-        expected_context = {
-            'replier_name': self.user_3.username,
-            'post_title': self.thread.title,
-            'author_name': 'your',
-            'course_name': self.course.display_name,
-        }
-        self.assertDictEqual(args_comment.context, expected_context)
-        self.assertEqual(
-            args_comment.content_url,
-            _get_mfe_url(self.course.id, self.thread.id)
-        )
-        self.assertEqual(args_comment.app_name, 'discussion')
-
-
-class TestSendCommentNotification(ForumsEnableMixin, CommentsServiceMockMixin, ModuleStoreTestCase):
-    """
-    Test case to send new_comment notification
-    """
-    def setUp(self):
-        super().setUp()
-        httpretty.reset()
-        httpretty.enable()
-
-        self.course = CourseFactory.create()
-        self.user_1 = UserFactory.create()
-        self.user_2 = UserFactory.create()
-
-    def test_new_comment_notification(self):
-        handler = Mock()
-        USER_NOTIFICATION_REQUESTED.connect(handler)
-
-        thread = ThreadMock(thread_id=1, creator=self.user_1, title='test thread')
-        response = ThreadMock(thread_id=2, creator=self.user_2, title='test response')
-        self.register_get_comment_response({
-            'id': response.id,
-            'thread_id': 'abc',
-            'user_id': response.user_id
-        })
-        send_response_notifications(thread, self.course, self.user_2, parent_id=response.id)
-        handler.assert_called_once()
-        context = handler.call_args[1]['notification_data'].context
-        self.assertEqual(context['author_name'], 'their')
 
 
 @ddt.ddt
