+++ lms/djangoapps/teams/tests/test_models.py
@@ -6,7 +6,7 @@ Tests for the teams API at the HTTP request level.
 import itertools
 from contextlib import contextmanager
 from datetime import datetime
-from unittest.mock import Mock
+from unittest.mock import MagicMock, patch
 
 import ddt
 import pytest
@@ -270,7 +270,7 @@ class TeamSignalsTest(EventTestMixin, SharedModuleStoreTestCase):
         """Create a mock comment service object with the given context."""
         if user is None:
             user = self.user
-        return Mock(
+        return MagicMock(
             user_id=user.id,
             commentable_id=self.DISCUSSION_TOPIC_ID,
             context=context,
@@ -318,8 +318,10 @@ class TeamSignalsTest(EventTestMixin, SharedModuleStoreTestCase):
         (user, should_update) = user_should_update
         with self.assert_last_activity_updated(should_update):
             user = getattr(self, user)
-            signal = self.SIGNALS[signal_name]
-            signal.send(sender=None, user=user, post=self.mock_comment())
+            with patch('lms.djangoapps.discussion.rest_api.tasks.send_response_notifications.apply_async'):
+                with patch('lms.djangoapps.discussion.rest_api.tasks.send_thread_created_notification.apply_async'):
+                    signal = self.SIGNALS[signal_name]
+                    signal.send(sender=None, user=user, post=self.mock_comment())
 
     @ddt.data('thread_voted', 'comment_voted')
     def test_vote_others_post(self, signal_name):
@@ -335,5 +337,7 @@ class TeamSignalsTest(EventTestMixin, SharedModuleStoreTestCase):
         place in discussions outside of a team.
         """
         with self.assert_last_activity_updated(False):
-            signal = self.SIGNALS[signal_name]
-            signal.send(sender=None, user=self.user, post=self.mock_comment(context='course'))
+            with patch('lms.djangoapps.discussion.rest_api.tasks.send_response_notifications.apply_async'):
+                with patch('lms.djangoapps.discussion.rest_api.tasks.send_thread_created_notification.apply_async'):
+                    signal = self.SIGNALS[signal_name]
+                    signal.send(sender=None, user=self.user, post=self.mock_comment(context='course'))
